+++ app/src/main/java/com/yesferal/hornsapp/app/presentation/base/BaseContract.kt
@@ -1,4 +1,4 @@
-package com.yesferal.hornsapp.app.presentation.common.base
+package com.yesferal.hornsapp.app.presentation.base
 
 interface BaseContract {
     interface View
