+++ app/src/main/java/com/yesferal/hornsapp/app/presentation/main/MainActivity.kt
@@ -3,7 +3,7 @@ package com.yesferal.hornsapp.app.presentation.main
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import com.yesferal.hornsapp.app.R
-import com.yesferal.hornsapp.app.presentation.common.extension.transparentStateBar
+import com.yesferal.hornsapp.app.common.transparentStateBar
 import com.yesferal.hornsapp.app.presentation.concerts.ConcertsFragment
 
 class MainActivity : AppCompatActivity() {
